# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "App Directory API",
    "description": "AI-First Application Directory \u2014 agents discover, submit, and rate AI-native services and tools",
    "version": "0.10.0",
    "x-event-types": {
      "description": "SSE event types emitted on /events/stream",
      "types": [
        "app.submitted",
        "app.approved",
        "app.rejected",
        "app.deprecated",
        "app.undeprecated",
        "app.updated",
        "app.deleted",
        "review.submitted",
        "health.checked",
        "warning"
      ]
    }
  },
  "servers": [
    {
      "url": "http://localhost:8002/api/v1"
    }
  ],
  "paths": {
...
    "/apps": {
      "get": {
        "summary": "List apps",
        "operationId": "listApps",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "protocol",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "approved"
            }
          },
          {
            "name": "featured",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter to featured apps only"
          },
          {
            "name": "verified",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter to verified apps only"
          },
          {
            "name": "health",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "healthy",
                "unhealthy",
                "unreachable",
                "unknown"
              ]
            },
            "description": "Filter by health check status"
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "newest",
                "rating",
                "name",
                "oldest"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of apps"
          }
        }
      },
      "post": {
        "summary": "Submit a new app",
        "operationId": "submitApp",
        "security": [
          {
            "apiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAppRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "App submitted"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/apps/{id}": {
      "get": {
        "summary": "Get app by ID or slug",
        "operationId": "getApp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "App details"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "summary": "Update app",
        "description": "Update app details. Auth via edit token (?token= query param or X-Edit-Token header), API key (owner), or admin key. Edit tokens cannot change status, featured, or verified badges.",
        "operationId": "updateApp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "description": "Edit token returned on app creation",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {},
          {
            "apiKey": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "App updated"
          }
        }
      },
      "delete": {
        "summary": "Delete app",
        "operationId": "deleteApp",
        "description": "Delete app and all associated data. Auth via edit token (?token= query param or X-Edit-Token header), API key (owner), or admin key.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "description": "Edit token returned on app creation",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {},
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "App deleted"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "SubmitAppRequest": {
        "type": "object",
        "required": [
          "name",
          "short_description",
          "description",
          "author_name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "short_description": {
            "type": "string",
            "maxLength": 200
          },
          "description": {
            "type": "string"
          },
          "homepage_url": {
            "type": "string",
            "format": "uri"
          },
          "api_url": {
            "type": "string",
            "format": "uri"
          },
          "api_spec_url": {
            "type": "string",
            "format": "uri",
            "description": "URL to OpenAPI/GraphQL/gRPC spec"
          },
          "protocol": {
            "type": "string",
            "enum": [
              "rest",
              "graphql",
              "grpc",
              "mcp",
              "a2a",
              "websocket",
              "other"
            ],
            "default": "rest"
          },
          "category": {
            "type": "string",
            "enum": [
              "communication",
              "data",
              "developer-tools",
              "finance",
              "media",
              "productivity",
              "search",
              "security",
              "social",
              "ai-ml",
              "infrastructure",
              "other"
            ],
            "default": "other"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "logo_url": {
            "type": "string",
            "format": "uri"
          },
          "author_name": {
            "type": "string"
          },
          "author_url": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "UpdateAppRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "short_description": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "homepage_url": {
            "type": "string"
          },
          "api_url": {
            "type": "string"
          },
          "api_spec_url": {
            "type": "string"
          },
          "protocol": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "logo_url": {
            "type": "string"
          },
          "author_name": {
            "type": "string"
          },
          "author_url": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "deprecated"
            ],
            "description": "Admin only"
          },
          "is_featured": {
            "type": "boolean",
            "description": "Admin only \u2014 highlight app as featured"
          },
          "is_verified": {
            "type": "boolean",
            "description": "Admin only \u2014 mark app as verified/trusted"
          }
        }
      },
...
    },
...
  }
}